+++ db/migrate/20240504185723_add_conventional_comments_count_to_insights.rb
@@ -1,0 +1,6 @@
+class AddConventionalCommentsCountToInsights < ActiveRecord::Migration[7.1]
+  def change
+    add_column :insights, :conventional_comments_count, :integer, default: 0
+    add_column :repositories_insights, :conventional_comments_count, :integer, default: 0
+  end
+end